+++ src/Codec/FFmpeg/Scaler.hs
@@ -1,9 +1,12 @@
-{-# LANGUAGE FlexibleContexts, ScopedTypeVariables, UndecidableInstances #-}
+{-# LANGUAGE FlexibleContexts, FlexibleInstances, ScopedTypeVariables,
+             UndecidableInstances #-}
 module Codec.FFmpeg.Scaler where
 import Codec.FFmpeg.Common
 import Codec.FFmpeg.Enums
+import Codec.FFmpeg.Internal.Linear (V2(..))
 import Codec.FFmpeg.Types
 import Codec.Picture
+import Data.Maybe (fromMaybe)
 import qualified Data.Vector.Storable as V
 import Foreign.C.Types
 import Foreign.Marshal.Array (withArray)
@@ -14,13 +17,17 @@ data ImageInfo = ImageInfo { imgWidth  :: CInt
                            , imgHeight :: CInt
                            , imgFormat :: AVPixelFormat }
 
+-- | @swsInit srcInfo dstInfo alg@ initializations an 'SwsContext' to
+-- scale and convert from @srcInfo@ to @dstInfo@ using the algorithm
+-- @alg@ when scaling.
 swsInit :: ImageInfo -> ImageInfo -> SwsAlgorithm -> IO SwsContext
 swsInit = swsReset (SwsContext nullPtr)
 
 -- | Obtain a context for converting the source to destination
 -- format. If the given context is already configured for the required
 -- conversion, it is returned. Otherwise, the given context is freed
--- and a new, configured context is returned.
+-- and a new, configured context is returned. See 'swsInit' for a
+-- description of the arguments.
 swsReset :: SwsContext -> ImageInfo -> ImageInfo -> SwsAlgorithm
          -> IO SwsContext
 swsReset ctx src dst alg = sws_getCachedContext ctx
@@ -51,6 +58,15 @@ instance (Pixel a, Storable (PixelBaseComponent a))
                componentCount (undefined :: a)
           stride = fromIntegral $ imageWidth img * sz
   sliceHeight img k = k (fromIntegral $ imageHeight img)
+
+instance SwsCompatible (AVPixelFormat, V2 CInt, V.Vector CUChar) where
+  swsPlanes (_,_,p) k = V.unsafeWith p $ \ptr ->
+                          withArray (castPtr ptr : replicate 7 nullPtr) k
+  swsStrides (fmt, V2 w _, _) k = withArray (stride : replicate 7 0) k
+    where sz = fromMaybe (error $ "Unknown pixel stride for format "++show fmt)
+                         (avPixelStride fmt)
+          stride = w * fromIntegral sz
+  sliceHeight (_, V2 _ h, _) k = k h
 
 -- | Supplies a continuation with all components provided by the
 -- 'SwsCompatible' class.